requirements.txt
streamlit==1.36.0
streamlit-webrtc==0.47.7
opencv-python
mediapipe
av
aiortc
pandas
numpy
scikit-learn
matplotlib
seaborn
tensorflow
twilio
requests
beautifulsoup4
googlesearch-python
pywhatkit
instagrapi
instabot
paramiko
docker
python-dotenv
